InstaShare/requirements.txt
alabaster==0.7.12
awscli==1.16.128
Babel==2.6.0
boto3==1.9.118
botocore==1.12.118
certifi==2018.11.29
chardet==3.0.4
colorama==0.3.9
Django==2.1.7
django-cors-headers==2.5.2
djangorestframework==3.9.1
djangorestframework-simplejwt==3.3
docutils==0.14
idna==2.8
imagesize==1.1.0
Jinja2==2.10
jmespath==0.9.4
MarkupSafe==1.1.1
packaging==19.0
Pillow==5.4.1
pyasn1==0.4.5
Pygments==2.3.1
PyJWT==1.7.1
pyparsing==2.3.1
python-dateutil==2.8.0
pytz==2018.9
PyYAML==3.13
requests==2.21.0
rsa==3.4.2
s3transfer==0.2.0
six==1.12.0
snowballstemmer==1.2.1
Sphinx==1.8.4
sphinxcontrib-websupport==1.1.0
urllib3==1.24.1
